agent-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
langgraph-checkpoint-postgres>=3.0.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
pydantic>=2.10.0
python-dotenv>=1.0.0
